Adds Cameras namespace
