+++ Assets/Scripts/Cameras/CameraController.cs
@@ -1,0 +1,100 @@
+using Inputs;
+using Mirror;
+using UnityEngine;
+using UnityEngine.InputSystem;
+
+namespace Cameras
+
+{
+    public class CameraController : NetworkBehaviour
+    {
+        [SerializeField]
+        private Transform playerCameraTransform;
+
+        [SerializeField]
+        private float speed = 20f;
+
+        [SerializeField]
+        private float screenBorderThickness = 10f;
+
+        [SerializeField]
+        private Vector2 screenXLimits = Vector2.zero;
+
+        [SerializeField]
+        private Vector2 screenZLimits = Vector2.zero;
+
+        private Vector2 previousInput;
+
+        private Controls controls;
+
+        public override void OnStartAuthority()
+        {
+            playerCameraTransform.gameObject.SetActive(true);
+
+            controls = new Controls();
+
+            controls.Player.MoveCamera.performed += SetPreviousInput;
+            controls.Player.MoveCamera.canceled += SetPreviousInput;
+
+            controls.Enable();
+        }
+
+        [ClientCallback]
+        private void Update()
+        {
+            if (!hasAuthority || !Application.isFocused)
+            {
+                return;
+            }
+
+            UpdateCameraPosition();
+        }
+
+        private void UpdateCameraPosition()
+        {
+            Vector3 position = playerCameraTransform.position;
+
+            if (previousInput == Vector2.zero)
+            {
+                Vector3 cursorMovement = Vector3.zero;
+
+                Vector2 cursorPosition = Mouse.current.position.ReadValue();
+
+                if (cursorPosition.y >= Screen.height - screenBorderThickness)
+                {
+                    cursorMovement.z += 1;
+                }
+                else if (cursorPosition.y <= screenBorderThickness)
+                {
+                    cursorMovement.z -= 1;
+                }
+
+                if (cursorPosition.x >= Screen.width - screenBorderThickness)
+                {
+                    cursorMovement.x += 1;
+                }
+                else if (cursorPosition.x <= screenBorderThickness)
+                {
+                    cursorMovement.x -= 1;
+                }
+
+                position += cursorMovement.normalized * (speed * Time.deltaTime);
+            }
+            else
+            {
+                position +=
+                    new Vector3(previousInput.x, 0f, previousInput.y) * (speed * Time.deltaTime);
+            }
+
+            position.x = Mathf.Clamp(position.x, screenXLimits.x, screenXLimits.y);
+            position.z = Mathf.Clamp(position.z, screenXLimits.x, screenXLimits.y);
+
+            playerCameraTransform.position = position;
+        }
+
+        private void SetPreviousInput(InputAction.CallbackContext ctx)
+        {
+            previousInput = ctx.ReadValue<Vector2>();
+        }
+    }
+}
