+++ ProjectSettings/RiderScriptEditorPersistedState.asset
@@ -12,4 +12,4 @@ MonoBehaviour:
   m_Script: {fileID: 0}
   m_Name: 
   m_EditorClassIdentifier: Unity.Rider.Editor:Packages.Rider.Editor:RiderScriptEditorPersistedState
-  lastWriteTicks: -8585462585631137844
+  lastWriteTicks: -8585462579205798443
